requirements.txt
streamlit==1.37.0
google-generativeai
streamlit_mic_recorder==0.0.6
